api/requirements.txt
django==3.1.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
django-reversion==3.0.8
django-cors-headers==3.5.0

requests
beautifulsoup4

flake8==3.8.4
mypy==0.790
pyflakes==2.2.0
pylint==2.6.0
pycodestyle==2.6.0
ipython==7.19.0
django_extensions==3.0.9
